Add nanosecond timestamp precision
because we don't measure server-side, for our experiments at the moment, we can actually provide extremely precise round trip info. We're still losing some precision by logging, quite a bit actually, but on average, it's pretty okay. The amount of gain is large enough that I don't have any reason to suspect it's a statistical inaccuracy.