Shell of cabling is finished
This adds the basic machinery for listening to gamepads at 256hz. We only intend to transmit at about 120hz, but we have to poll, and we want to catch input as early as we can. I'd like to go up to 512, but I just don't know how fast this stuff actually can be. We'll see. The xb1+ controllers really operate between 120hz and 160hz, but we're not clock synced, so we're oversampling. dual shocks operate at up to 256hz by some accounts, but I haven't run those numbers down yet.