Add a force timestamp method, maybe fix a crash, fix DSCP
I can now confirm that DSCP has a definitive effect, and that if you want to get good latency variance prevention out of bristlecone, a variety of DSCP signifiers IS the best way to do it on wired. On wifi, you always get the benefit. I'm still dialing in what the right three classes are, but this at least gives interesting behavior that's a good jumping off point.