Added justification and a workable DSCP load-out
Bullocks. Windows doesn't seem to allow AHB. I have validated that packets arrive out of order, that bristlecone reduces latency over the wire due to packet reordering, and that it provides added packet resilience. The incidence of this on wired connections isn't huge, but it does add up fast over a 5 minute match.