Adjust includes to eliminate excessive type exposure
This helps permanently resolve most of the warnings and potential hazards involved in including the gameinput library this way. However, we may still wish to ensure the DLL ships along with in the expected way. Hard to say. In any case, we now stage, prep, and package correctly in so far as I can tell. Tomorrow, I can write the actual control loop and queuing. Pretty pumped.